change package name to com.eight.nivadeus
