+++ build.gradle
@@ -20,7 +20,7 @@ apply plugin: 'org.parchmentmc.librarian.forgegradle'
 
 
 version = '0.01-1.18.2'
-group = 'com.happy.nivadeus' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+group = 'com.eight.nivadeus' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'nivadeus'
 
 // Mojang ships Java 17 to end users in 1.18+, so your mod should target Java 17.
