+++ src/main/java/com/eight/nivadeus/manasystem/data/Mana.java
@@ -1,0 +1,18 @@
+package com.eight.nivadeus.manasystem.data;
+
+public class Mana {
+
+	private int mana;
+
+	public Mana(int mana) {
+		this.mana = mana;
+	}
+
+	public int getMana() {
+		return mana;
+	}
+
+	public void setMana(int mana) {
+		this.mana = mana;
+	}
+}
