+++ src/main/java/com/eight/nivadeus/setup/ClientSetup.java
@@ -1,0 +1,27 @@
+package com.eight.nivadeus.setup;
+
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.gui.ForgeIngameGui;
+import net.minecraftforge.client.gui.OverlayRegistry;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+
+import static net.minecraftforge.client.gui.ForgeIngameGui.HOTBAR_ELEMENT;
+
+import com.eight.nivadeus.Nivadeus;
+import com.eight.nivadeus.manasystem.client.KeyBindings;
+import com.eight.nivadeus.manasystem.client.KeyInputHandler;
+import com.eight.nivadeus.manasystem.client.ManaOverlay;
+
+@EventBusSubscriber(modid = Nivadeus.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
+public class ClientSetup {
+	
+	public static void init(FMLClientSetupEvent event) {
+        MinecraftForge.EVENT_BUS.addListener(KeyInputHandler::onKeyInput);
+		KeyBindings.init();
+		OverlayRegistry.registerOverlayAbove(HOTBAR_ELEMENT, "name", ManaOverlay.HUD_MANA);
+	}
+
+}
