--- src/main/java/com/happy/nivadeus/manasystem/data/Mana.java
@@ -1,18 +1,0 @@
-package com.happy.nivadeus.manasystem.data;
-
-public class Mana {
-
-	private int mana;
-
-	public Mana(int mana) {
-		this.mana = mana;
-	}
-
-	public int getMana() {
-		return mana;
-	}
-
-	public void setMana(int mana) {
-		this.mana = mana;
-	}
-}
