--- src/main/java/com/happy/nivadeus/setup/Config.java
@@ -1,33 +1,0 @@
-package com.happy.nivadeus.setup;
-
-import com.happy.nivadeus.manasystem.ManaConfig;
-
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.config.ModConfig;
-
-public class Config {
-
-	public static void register() {
-		registerClientConfigs();
-		registerCommonConfigs();
-		registerServerConfigs();
-	}
-	
-	private static void registerClientConfigs() {
-        ForgeConfigSpec.Builder CLIENT_BUILDER = new ForgeConfigSpec.Builder();
-        ManaConfig.registerClientConfig(CLIENT_BUILDER);
-        ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, CLIENT_BUILDER.build());
-    }
-
-    private static void registerCommonConfigs() {
-        ForgeConfigSpec.Builder COMMON_BUILDER = new ForgeConfigSpec.Builder();
-        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, COMMON_BUILDER.build());
-    }
-
-    private static void registerServerConfigs() {
-        ForgeConfigSpec.Builder SERVER_BUILDER = new ForgeConfigSpec.Builder();
-        ManaConfig.registerServerConfig(SERVER_BUILDER);
-        ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, SERVER_BUILDER.build());
-    }
-}
